+++ src/generateSnippet.ts
@@ -7,7 +7,9 @@ export function generateSnippet(snippetObject, allSnippets) {
   const filename = path.parse(snippetObject.docFile).base;
   // Feature? define in config the format of the prefix
   const resourceName = filename.split(".")[0];
-  const fullResourceName = `${snippetObject.provider}-${snippetObject.resourceType}-${resourceName}`;
+  const fullResourceName = `${snippetObject.provider}-${
+    snippetObject.resourceType
+  }-${resourceName}`;
   snippetObject.snippet[fullResourceName] = {
     description: `Defines ${resourceName}`,
     prefix: `tf-${fullResourceName}`,