+++ src/setup.ts
@@ -5,5 +5,5 @@ export function setupWorkspace() {
   const tmpobj = tmp.dirSync();
   const tmpDir = tmpobj.name;
   console.log(`tmp dir will be ${tmpDir}`);
-  return tmpDir
+  return tmpDir;
 }